+++ admin/tabs/cluster/index.ts
@@ -1,0 +1,2 @@
+export * from './instances/InstancesTab';
+export * from './objects/ClusterObjectsPanel';