+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 75.0-SNAPSHOT - Unreleased
 
+### 💥 Breaking Changes
+
+* Renamed `TabModel.containerModel` prop to `parentContainerModel` for clarity.
+
 ### 🎁 New Features
 
 * New property `GridModel.expandToLevel` governs the expansion state of tree and grouped grids.
@@ -12,6 +16,8 @@
       property for your grid.
 * Added new `GroupingChooserModel.sortDimensions` config - can be set to false to respect the order
   in which dimensions are provided to the model.
+* Added new `childTabs` property of type `TabContainerProps` to `TabConfig`. This enables tabs to render
+  their own nested tab containers as content.
 
 ### 🐞 Bug Fixes